requirements.txt
summarizer
bert-extractive-summarizer
newspaper3k
spacy
transformers